Move dashboard storage configuration from webapp to storage
