+++ server/storage/src/main/java/org/bithon/server/storage/StorageModuleAutoConfiguration.java
@@ -16,8 +16,12 @@
 
 package org.bithon.server.storage;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.bithon.component.commons.utils.StringUtils;
 import org.bithon.server.storage.common.provider.StorageProviderManager;
+import org.bithon.server.storage.dashboard.DashboardStorageConfig;
+import org.bithon.server.storage.dashboard.IDashboardStorage;
 import org.bithon.server.storage.datasource.SchemaManager;
 import org.bithon.server.storage.event.EventStorageConfig;
 import org.bithon.server.storage.event.EventTableSchema;
@@ -187,5 +191,49 @@ public class StorageModuleAutoConfiguration {
             initializer.initialize(manager);
         }
         return manager;
+    }
+
+    @Bean
+    @ConditionalOnProperty(value = "bithon.storage.dashboard.enabled", havingValue = "true")
+    public IDashboardStorage dashboardStorage(ObjectMapper om,
+                                              StorageProviderManager storageProviderManager,
+                                              DashboardStorageConfig storageConfig) throws IOException {
+        String providerName = StringUtils.isEmpty(storageConfig.getProvider()) ? storageConfig.getType() : storageConfig.getProvider();
+        InvalidConfigurationException.throwIf(!StringUtils.hasText(providerName),
+                                              "[%s] is not properly configured to enable the storage for the dashboard module.",
+                                              storageConfig.getClass(),
+                                              "provider");
+
+        // create storage
+        IDashboardStorage storage = storageProviderManager.createStorage(providerName, IDashboardStorage.class);
+        storage.initialize();
+
+        // load or update schemas
+        try {
+            Resource[] resources = new PathMatchingResourcePatternResolver()
+                .getResources("classpath:/dashboard/*.json");
+            for (Resource resource : resources) {
+
+                JsonNode dashboard = om.readTree(resource.getInputStream());
+                JsonNode nameNode = dashboard.get("name");
+                if (nameNode == null) {
+                    throw new RuntimeException(StringUtils.format("dashboard [%s] miss the name property", resource.getFilename()));
+                }
+
+                String name = nameNode.asText();
+                if (StringUtils.isEmpty(name)) {
+                    throw new RuntimeException(StringUtils.format("dashboard [%s] has empty name property", resource.getFilename()));
+                }
+
+                // deserialize and then serialize again to compact the json string
+                String payload = om.writeValueAsString(dashboard);
+
+                storage.putIfNotExist(nameNode.asText(), payload);
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        return storage;
     }
 }
