+++ server/storage/src/main/resources/dashboard/kafka-consumer-metrics.json
