+++ server/storage/src/main/resources/dashboard/kafka-network-metrics.json
