+++ server/storage/src/main/resources/dashboard/kafka-producer-metrics.json
