+++ server/storage/src/main/resources/dashboard/mongodb-metrics.json
