+++ server/storage/src/main/resources/dashboard/redis-metrics.json
