+++ server/storage/src/main/resources/dashboard/web-server-metrics.json
