+++ server/web-service/src/main/java/org/bithon/server/web/service/dashboard/DashboardManager.java
@@ -23,8 +23,8 @@ import org.bithon.component.commons.concurrency.NamedThreadFactory;
 import org.bithon.component.commons.concurrency.ScheduledExecutorServiceFactor;
 import org.bithon.component.commons.time.DateTime;
 import org.bithon.server.commons.time.TimeSpan;
-import org.bithon.server.storage.web.Dashboard;
-import org.bithon.server.storage.web.IDashboardStorage;
+import org.bithon.server.storage.dashboard.Dashboard;
+import org.bithon.server.storage.dashboard.IDashboardStorage;
 import org.bithon.server.web.service.WebServiceModuleEnabler;
 import org.springframework.context.SmartLifecycle;
 import org.springframework.context.annotation.Conditional;
